[CALCITE-3479] Stack overflow error thrown when running join queryi (Xiening Dai)
With fix of CALCITE-2166, we recalculate the RelNode cost within a subset when
its best cost is increased, in order to find the new best RelNode. But this
could lead to cyclic reference between best RelNodes and then fail to generate
a plan.

In this change, we remove the recalcuation part to mitigate this problem.

Close #1567

(cherry picked from commit 2ab44bff3a086acae7b38dcff264060544860147)